Hexagon: HVX sysemu, tests, fixes

Added HVX support for system emulation: bind SSR.XA to the associated
HVX context.

Endian bugfixes for HVX.

Add tests for improved coverage.

Improvements in the simple packet short-circuit logic - for handling
multiple reg-writes to GPRs, System, Guest, Vector, Vec-Pred registers.

# -----BEGIN PGP SIGNATURE-----
#
# iQIzBAABCgAdFiEEPWaq5HRZSCTIjOD4GlSvuOVkbDIFAmq11C0ACgkQGlSvuOVk
# bDJ8iw/8DAoZzi6Ph7GYOfkjZX++JxYfmjKjAkL1pbj2Ht8a2LkWGmch9hy9zm+i
# rarfkLlaz4HyMk2N/HeZHxkRmGLSLfoM9zR80jkwnHNw9E4AAFzPBMCJKeknmitR
# rwUCs9AVbfH3gURUty1GWybmjCXgxjk9N9S9Z22Sll/bk2XQw+IeRbW6ho5PwKl0
# lnJWm+cDjlRoi4E3auB8uvWRUMPkhDleN7CHDZEJSTFiNG1drRCt8+BUO24CaNu0
# Z8LeDyNkIUxKLtbb04gDmY3gh0E+LHxZ7iSH1vQf1hhZPNDG5vVAap0ZTgy1XjVY
# rZFh9Jo/uw2rncxi96GWIlA0NQOzXvNVYmHZuHINBRnsgFwlhA4gspYgGEjHoY1f
# tr62OSp7yPUpDnaV2RK93mRPXCRaytaJ8FF/vKPKSXd4hEuMcQHLdJk9/BeAZaJs
# CWPn5Fc206rUqq7WAV+qkAu/DttoWc+lQ/NQEH0MGxbMfmHTf9zpMJJALRyFwzAa
# lGbt44a5JnMi7geUoXY4wusVDFQCzzD4D0oWebSCMB3cYluZ7WVb0OGt+eCDjrDl
# fhbrtoTh5N/+KqhKqt+8oEy0QFRERfM3ebN/uWg6n3bGDyl6CrIa9kMdFurIz7nA
# LxxniNyGZ0cTlu2cbTSg+A5ocVQcupq5hgpQjKa+uUoIBJewk2o=
# =+KEV
# -----END PGP SIGNATURE-----
# gpg: Signature made Thu 24 Sep 2026 06:53:49 PM PDT
# gpg:                using RSA key 3D66AAE474594824C88CE0F81A54AFB8E5646C32
# gpg: Good signature from "Brian Cain (OSS Qualcomm) <brian.cain@oss.qualcomm.com>" [unknown]
# gpg:                 aka "Brian Cain (QUIC) <quic_bcain@quicinc.com>" [unknown]
# gpg:                 aka "Brian Cain <bcain@kernel.org>" [unknown]
# gpg:                 aka "Brian Cain (QuIC) <bcain@quicinc.com>" [unknown]
# gpg:                 aka "Brian Cain (CAF) <bcain@codeaurora.org>" [unknown]
# gpg:                 aka "bcain" [unknown]
# gpg: WARNING: This key is not certified with a trusted signature!
# gpg:          There is no indication that the signature belongs to the owner.
# Primary key fingerprint: 6350 20F9 67A7 7164 79EF  49E0 175C 464E 541B 6D47
#      Subkey fingerprint: 3D66 AAE4 7459 4824 C88C  E0F8 1A54 AFB8 E564 6C32

* tag 'pull-hex-20260924' of https://github.com/qualcomm/qemu: (40 commits)
  target/hexagon: rename LEN to OFFS_REG_END in VTCM macros
  tests/tcg/hexagon: check the two HVX predicate build paths agree
  target/hexagon: fix HVX predicate save size for histogram ops
  tests/tcg/hexagon: add vgather/vscatter region-length tests
  target/hexagon: fix HVX scatter/gather region-length check
  target/hexagon: fix HVX big-endian byte access
  tests/functional/hexagon: add HVX tests
  target/hexagon: share HVX contexts among CPUs
  target/hexagon: prepare for separate HVX contexts
  target/hexagon: raise an exception for HVX with SSR:XE clear
  tcg: split out _var forms of gvec saturating/mul/shift-immediate ops
  tcg: add base arguments to the gvec 2s, dup_i32 and cmp expanders
  tcg: split out _var forms of gvec logical, abs and min/max ops
  tests/tcg/hexagon: add decbin MPS test case
  tests/tcg/hexagon: add scalar+HVX and masked store tests
  tests/tcg/hexagon: add rounding conv, sffma:lib, and sfinvsqrta
  tests/tcg/hexagon: add predicate register transfer test
  tests/tcg/hexagon: add unconditional store-immediate tests
  tests/tcg/hexagon: add dfmpyhh
  tests/tcg/hexagon: add sfrecipa edge case tests
  ...

+++ tests/tcg/hexagon/circ.c
@@ -97,9 +97,9 @@ INIT(dbuf, NDOBLS)
  */
 static int32_t build_mreg(int32_t inc, int32_t K, int32_t len)
 {
-    return ((inc & 0x780) << 21) |
-           ((K & 0xf) << 24) |
-           ((inc & 0x7f) << 17) |
+    return ((uint32_t)(inc & 0x780) << 21) |
+           ((uint32_t)(K & 0xf) << 24) |
+           ((uint32_t)(inc & 0x7f) << 17) |
            (len & 0x1ffff);
 }
